refactor: move integrations to pkg/integrations folder
growthbook: move growthbook

mixpanel: move mixpanel

sentry: move sentry
