+++ src/pkg/integrations/mixpanel/mixpanel.provider.tsx
